% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRanking.R
\name{TaskRanking}
\alias{TaskRanking}
\title{TaskRanking Class}
\format{
An R6 class inheriting from \code{TaskDecision}.
}
\description{
Task class for ranking problems in Multi-Criteria Decision Analysis.
Inherits from \code{\link{TaskDecision}} and automatically sets task type to "ranking".
}
\details{
This is a convenience class for ranking tasks. It automatically sets
\code{type = "ranking"} when initializing. All other functionality is
inherited from \code{TaskDecision}.
}
\section{Methods}{

\describe{
\item{\code{initialize(..., type)}}{Create a new TaskDecision object}
\item{\code{validate()}}{Validate all data and constraints}
\item{\code{get_perf(alt_name, crit_name)}}{Get performance matrix or single value}
\item{\code{n_alt()}}{Get number of alternatives}
\item{\code{n_crit()}}{Get number of criteria}
\item{\code{alt_names()}}{Get alternative names}
\item{\code{crit_names()}}{Get criterion names}
}
}

\examples{
\dontrun{
# Load toy dataset
data(cars, package = "mcdaHub")

# Create a ranking task
task <- TaskRanking$new(
  alt = cars$alt,
  crit = cars$crit,
  perf = cars$perf,
  weight = cars$weight,
  direction = cars$direction
)
}


## ------------------------------------------------
## Method `TaskRanking$new`
## ------------------------------------------------

\dontrun{
data(cars, package = "mcdaHub")
task <- TaskRanking$new(
  alt = cars$alt,
  crit = cars$crit,
  perf = cars$perf,
  weight = cars$weight,
  direction = cars$direction
)
}
}
\seealso{
\code{\link{TaskDecision}} for base class documentation.
}
\section{Super class}{
\code{\link[mcdaHub:TaskDecision]{mcdaHub::TaskDecision}} -> \code{TaskRanking}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskRanking-new}{\code{TaskRanking$new()}}
\item \href{#method-TaskRanking-print}{\code{TaskRanking$print()}}
\item \href{#method-TaskRanking-clone}{\code{TaskRanking$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="alt_names"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-alt_names'><code>mcdaHub::TaskDecision$alt_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="crit_names"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-crit_names'><code>mcdaHub::TaskDecision$crit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="get_perf"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-get_perf'><code>mcdaHub::TaskDecision$get_perf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="n_alt"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-n_alt'><code>mcdaHub::TaskDecision$n_alt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="n_crit"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-n_crit'><code>mcdaHub::TaskDecision$n_crit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mcdaHub" data-topic="TaskDecision" data-id="validate"><a href='../../mcdaHub/html/TaskDecision.html#method-TaskDecision-validate'><code>mcdaHub::TaskDecision$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskRanking-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskRanking-new}{}}}
\subsection{Method \code{new()}}{
Create a new TaskRanking object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskRanking$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Individual fields as named arguments.
See \code{\link{TaskDecision}} for available fields.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A TaskRanking object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data(cars, package = "mcdaHub")
task <- TaskRanking$new(
  alt = cars$alt,
  crit = cars$crit,
  perf = cars$perf,
  weight = cars$weight,
  direction = cars$direction
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskRanking-print"></a>}}
\if{latex}{\out{\hypertarget{method-TaskRanking-print}{}}}
\subsection{Method \code{print()}}{
Print task information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskRanking$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskRanking-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskRanking-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskRanking$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
